pyproject.toml
[project]
name = "ml-portfolio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.1",
    "flask-mail>=0.10.0",
    "flask-wtf>=1.2.2",
    "gunicorn>=23.0.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "ruff>=0.11.12",
]
